# packages/backend/src/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/verify-totp": {
      "post": {
        "operationId": "verifyTotp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTotpBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/username": {
      "patch": {
        "operationId": "changeUsername",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUsernameBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/password": {
      "patch": {
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/totp/get-uri": {
      "patch": {
        "operationId": "getTotpUri",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTotpUriBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTotpUriDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/totp/setup": {
      "patch": {
        "operationId": "setupTotp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupTotpBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/totp/disable": {
      "patch": {
        "operationId": "disableTotp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableTotpBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "delete": {
        "operationId": "cancelResetPassword",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      },
      "get": {
        "operationId": "checkResetPasswordRequest",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResetPasswordRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/backups/{urn}/backup": {
      "post": {
        "operationId": "backupApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Backups"
        ]
      }
    },
    "/api/backups/{urn}/restore": {
      "post": {
        "operationId": "restoreAppBackup",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreAppBackupDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Backups"
        ]
      }
    },
    "/api/backups/{urn}": {
      "get": {
        "operationId": "getAppBackups",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string",
                  "pattern": "^(?:(?!^-0$)-?(?:(?:0|[1-9]\\d*)))$"
                }
              ],
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string",
                  "pattern": "^(?:(?!^-0$)-?(?:(?:0|[1-9]\\d*)))$"
                }
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAppBackupsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Backups"
        ]
      },
      "delete": {
        "operationId": "deleteAppBackup",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAppBackupBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Backups"
        ]
      }
    },
...
    "/api/links": {
      "get": {
        "operationId": "getLinks",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinksDto"
                }
              }
            }
          }
        },
        "tags": [
          "Links"
        ]
      },
      "post": {
        "operationId": "createLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Links"
        ]
      }
    },
    "/api/links/{id}": {
      "patch": {
        "operationId": "editLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditLinkBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Links"
        ]
      },
      "delete": {
        "operationId": "deleteLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Links"
        ]
      }
    },
...
  },
  "info": {
    "title": "Runtipi API",
    "description": "API specs for Runtipi",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LoginBody": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "totpSessionId": {
            "type": "string"
          }
        },
        "required": [
          "success"
        ]
      },
      "VerifyTotpBody": {
        "type": "object",
        "properties": {
          "totpCode": {
            "type": "string"
          },
          "totpSessionId": {
            "type": "string"
          }
        },
        "required": [
          "totpCode",
          "totpSessionId"
        ]
      },
      "RegisterBody": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "ChangeUsernameBody": {
        "type": "object",
        "properties": {
          "newUsername": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "newUsername",
          "password"
        ]
      },
      "ChangePasswordBody": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "GetTotpUriBody": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
      "GetTotpUriDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "uri": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "uri"
        ]
      },
      "SetupTotpBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "DisableTotpBody": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
      "ResetPasswordBody": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "newPassword"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "email",
          "success"
        ]
      },
      "CheckResetPasswordRequestDto": {
        "type": "object",
        "properties": {
          "isRequestPending": {
            "type": "boolean"
          }
        },
        "required": [
          "isRequestPending"
        ]
      },
...
      "BackupRequestDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "anyOf": [
              {
                "type": "string",
                "description": "a UUID",
                "format": "uuid",
                "pattern": "[\\da-f]{8}-[\\da-f]{4}-[1-8][\\da-f]{3}-[89ab][\\da-f]{3}-[\\da-f]{12}"
              },
              {
                "const": "00000000-0000-0000-0000-000000000000",
                "description": "a UUID",
                "format": "uuid"
              },
              {
                "const": "ffffffff-ffff-ffff-ffff-ffffffffffff",
                "description": "a UUID",
                "format": "uuid"
              }
            ],
            "description": "a UUID",
            "format": "uuid"
          }
        },
        "required": [
          "requestId"
        ]
      },
      "RestoreAppBackupDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          }
        },
        "required": [
          "filename"
        ]
      },
      "GetAppBackupsDto": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "number"
                },
                "id": {
                  "type": "string"
                },
                "size": {
                  "type": "number"
                }
              },
              "required": [
                "date",
                "id",
                "size"
              ]
            }
          },
          "lastPage": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "currentPage",
          "data",
          "lastPage",
          "total"
        ]
      },
      "DeleteAppBackupBodyDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          }
        },
        "required": [
          "filename"
        ]
      },
      "LinksDto": {
        "type": "object",
        "properties": {
          "links": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "description": {
                  "anyOf": [
                    {
                      "type": "string",
                      "maxLength": 50
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "iconUrl": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "uri"
                    },
                    {
                      "const": ""
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "id": {
                  "type": "number"
                },
                "isVisibleOnGuestDashboard": {
                  "type": "boolean"
                },
                "title": {
                  "type": "string",
                  "maxLength": 20,
                  "minLength": 1
                },
                "url": {
                  "type": "string",
                  "format": "uri"
                },
                "userId": {
                  "type": "number"
                }
              },
              "required": [
                "description",
                "iconUrl",
                "id",
                "isVisibleOnGuestDashboard",
                "title",
                "url",
                "userId"
              ]
            }
          }
        },
        "required": [
          "links"
        ]
      },
      "LinkBodyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "isVisibleOnGuestDashboard": {
            "type": "boolean",
            "default": false
          },
          "description": {
            "type": "string",
            "maxLength": 50
          },
          "iconUrl": {
            "anyOf": [
              {
                "type": "string",
                "format": "uri"
              }
            ]
          }
        },
        "required": [
          "title",
          "url"
        ]
      },
      "EditLinkBodyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string",
            "maxLength": 50
          },
          "iconUrl": {
            "anyOf": [
              {
                "type": "string",
                "format": "uri"
              }
            ]
          },
          "isVisibleOnGuestDashboard": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "url"
        ]
      },
...
    }
  }
}